requirements.txt
Django
Django-admin
Pillow
chardet
psycopg2
